% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_start_book.R
\name{q_start_book}
\alias{q_start_book}
\title{Start Quarto Book}
\usage{
q_start_book(author = "Lingyun Zhang", title = "My New Book")
}
\arguments{
\item{author}{author of the book, a character}

\item{title}{title of the book, a character}
}
\value{
TRUE, if successfully start a book
}
\description{
Start Quarto Book
}
