% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_book.R
\name{start_book}
\alias{start_book}
\title{Start Book}
\usage{
start_book(title = "My Book", author = "Lingyun Zhang")
}
\arguments{
\item{title}{book's title, a character}

\item{author}{book's author, a character}
}
\value{
TRUE, if successfully start a book
}
\description{
Start Book
}
