% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_tex.R
\name{start_tex}
\alias{start_tex}
\title{Title Start a Tex File}
\usage{
start_tex(tex_file_name = "my_paper.tex")
}
\arguments{
\item{tex_file_name}{a character}
}
\value{
TRUE, if succeed
}
\description{
Title Start a Tex File
}
